everything works well omg

+++ srcs/ft_llutoa_hex.c
@@ -6,13 +6,13 @@
 /*   By: pdal-mol <dolmalinn@gmail.com>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/11 16:44:21 by pdal-mol          #+#    #+#             */
-/*   Updated: 2021/11/16 10:55:11 by pdal-mol         ###   ########.fr       */
+/*   Updated: 2021/11/16 11:32:55 by pdal-mol         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libftprintf.h"
+#include "ft_printf.h"
 
-char	*ft_llutoa_hex(unsigned long long nb, char *base)
+char	*ft_llutoa_hex(unsigned long nb, char *base)
 {
 	char	*output;
 	size_t	i;
